+++ src/Core/Framework/DataAbstractionLayer/Search/AggregationResult/DefaultFacetExtension.php
@@ -33,7 +33,7 @@
 namespace Elio\ElioDataDiscovery\Core\Framework\DataAbstractionLayer\Search\AggregationResult;
 
 
-use Elio\ElioDataDiscovery\Core\Util\PropertyUtil;
+use Elio\ElioDataDiscovery\Core\Util\StringUtil;
 use Shopware\Core\Framework\Struct\Struct;
 
 /**
@@ -134,7 +134,7 @@ class DefaultFacetExtension extends Struct
     private static function sanitizeFacet(string $name): string
     {
         $parts = explode('.', $name);
-        $parts[count($parts) - 1] = PropertyUtil::encodeStringToUnicodeEscaped(end($parts));
+        $parts[count($parts) - 1] = StringUtil::encodeStringToUnicodeEscaped(end($parts));
         return str_replace('\\', '', implode('.', $parts));
     }
 }